+++ src/Pim/Bundle/EnrichBundle/Controller/GroupController.php
@@ -3,6 +3,7 @@
 namespace Pim\Bundle\EnrichBundle\Controller;
 
 use Oro\Bundle\SecurityBundle\Annotation\AclAncestor;
+use Pim\Bundle\CatalogBundle\Factory\GroupFactory;
 use Pim\Bundle\CatalogBundle\Model\GroupInterface;
 use Pim\Bundle\CatalogBundle\Entity\Group;
 use Pim\Bundle\CatalogBundle\Manager\GroupManager;
@@ -41,6 +42,9 @@ class GroupController extends AbstractController
     /** @var Form */
     protected $groupForm;
 
+    /** @var GroupFactory */
+    protected $groupFactory;
+
     /**
      * Constructor
      *
@@ -55,6 +59,7 @@ class GroupController extends AbstractController
      * @param GroupManager             $groupManager
      * @param HandlerInterface         $groupHandler
      * @param Form                     $groupForm
+     * @param GroupFactory             $groupFactory
      */
     public function __construct(
         Request $request,
@@ -67,7 +72,8 @@ class GroupController extends AbstractController
         EventDispatcherInterface $eventDispatcher,
         GroupManager $groupManager,
         HandlerInterface $groupHandler,
-        Form $groupForm
+        Form $groupForm,
+        GroupFactory $groupFactory
     ) {
         parent::__construct(
             $request,
@@ -83,6 +89,7 @@ class GroupController extends AbstractController
         $this->groupManager = $groupManager;
         $this->groupHandler = $groupHandler;
         $this->groupForm    = $groupForm;
+        $this->groupFactory = $groupFactory;
     }
 
     /**
@@ -115,8 +122,7 @@ class GroupController extends AbstractController
             return $this->redirectToRoute('pim_enrich_group_index');
         }
 
-        // TODO : inject class or use a factory
-        $group = new Group();
+        $group = $this->groupFactory->createGroup();
 
         if ($this->groupHandler->process($group)) {
             $this->addFlash('success', 'flash.group.created');